Backend/requirements.txt
openparse==0.7.0
python-dotenv==1.0.1
llama-index==0.12.9